+++ pystiche_papers/sanakoyeu_et_al_2018/__init__.py
@@ -1,5 +1,6 @@
 from ._augmentation import *
 from ._data import *
+from ._discriminator import *
 from ._loss import *
 from ._modules import *
 from ._nst import *
